Scaffold faithfulness eval: infrastructure, ablation flags, UX fixes, pilot runs
- Add eval infrastructure: contamination audit, clean-room isolated runner,
  stimulus driver (run_isolated.py, run_stimulus.py, contamination_audit.py)
- Add ablation flags (A=frozen-spec, B=provenance-metrics, C=split-disjointness)
  for leave-one-out discipline measurement
- Add free-agent (C0) harness for RQ1 baseline comparison
- Fix agent path resolution: _resolve_agent_path handles both run-dir-relative
  and repo-root-relative paths without doubling (read_file, write_file, edit_file)
- Fix compute_and_commit_metric to accept file paths for large inputs
- Fix model loading to default to fp32 (was bf16, causing quantized logit-diffs)
- Add complement-ablation faithfulness docs to activation-patching skill
- Strip spec provenance metadata to pass contamination audit
- Include S2 IOI pilot runs (C0, full) with scoring artifacts
- Add eval specs for S1 (shutdown), S2 (IOI), S3 (surprise)
- Add tests for ablation flags, free-agent harness, metric contracts

+++ src/autointerp/tools/model.py
@@ -64,7 +64,7 @@ class ModelHandle:
     model: Any
     tokenizer: Any
     device: str = "cuda"
-    dtype: torch.dtype = torch.bfloat16
+    dtype: torch.dtype = torch.float32
     model_type: str = "unknown"
 
     @property
@@ -219,7 +219,7 @@ class ModelHandle:
 def load_model(
     model_name: str,
     device: str = "cuda",
-    dtype: torch.dtype = torch.bfloat16,
+    dtype: torch.dtype = torch.float32,
     quantization: Optional[str] = None,
     trust_remote_code: bool = True,
 ) -> ModelHandle:
